+++ sedat/CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 0.2.3
+
+- Added [closeAll] method to [AbstractSecureDatabaseConfigurer].
+- Added [dispose] method to [BaseRepository].
+- Fixed [stream] issues in [BaseRepository].